0.3 Update.
The 0.3 update represents a major overhaul of one of the games core mechanics.

In updates 0.2 and below, the game played out in a linear way, where the game had 3 levels.

In update 0.3, the objective is to now survive for 1 minute while collecting the most points.

This makes the points system more useful, as it now challenges the player to collect as many points as they can find from the collectibles currently available.

I am working on a new player movement script which should be implemented in a 0.4 update, after which the game will move onto the beta phase.